+++ web/oauth.go
@@ -21,7 +21,6 @@ import (
 	"github.com/G-Node/gin-auth/conf"
 	"github.com/G-Node/gin-auth/data"
 	"github.com/G-Node/gin-auth/util"
-	"github.com/dchest/captcha"
 	"github.com/gorilla/mux"
 )
 
@@ -733,425 +732,4 @@ func Validate(w http.ResponseWriter, r *http.Request) {
 	w.Header().Add("Content-Type", "application/json")
 	enc := json.NewEncoder(w)
 	enc.Encode(response)
-}
-
-type validateAccount struct {
-	*data.Account
-	*util.ValidationError
-	CaptchaId      string
-	CaptchaResolve string
-}
-
-// RegistrationPage displays entry fields required for the creation of a new gin account
-func RegistrationPage(w http.ResponseWriter, r *http.Request) {
-	valAccount := &validateAccount{}
-	valAccount.Account = &data.Account{}
-	valAccount.ValidationError = &util.ValidationError{}
-	valAccount.CaptchaId = captcha.New()
-
-	tmpl := conf.MakeTemplate("registration.html")
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-	err := tmpl.ExecuteTemplate(w, "layout", valAccount)
-	if err != nil {
-		panic(err)
-	}
-}
-
-type passwordData struct {
-	Password        string
-	PasswordControl string
-}
-
-type registration struct {
-	verifyCaptcha func(string, string) bool
-}
-
-// RegistrationHandler provides an http handler for account registration.
-func RegistrationHandler(f func(string, string) bool) http.Handler {
-	rh := &registration{verifyCaptcha: f}
-	return rh
-}
-
-// The http handler of the registration class parses user entries for a new account. It will redirect back to the
-// entry form, if input is invalid. If the input is correct, it will create a new account,
-// send an e-mail with an activation link and redirect to the the registered page.
-func (rh *registration) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	tmpl := conf.MakeTemplate("registration.html")
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-
-	account := &data.Account{}
-	pw := &passwordData{}
-
-	err := util.ReadFormIntoStruct(r, account, true)
-	if err != nil {
-		PrintErrorHTML(w, r, err, http.StatusInternalServerError)
-		return
-	}
-
-	err = util.ReadFormIntoStruct(r, pw, true)
-	if err != nil {
-		PrintErrorHTML(w, r, err, http.StatusInternalServerError)
-		return
-	}
-
-	valAccount := &validateAccount{}
-	valAccount.ValidationError = &util.ValidationError{}
-	valAccount.Account = account
-
-	if r.Form.Encode() == "" {
-		valAccount.Message = "Please add all required fields (*)"
-		err := tmpl.ExecuteTemplate(w, "layout", valAccount)
-		if err != nil {
-			panic(err)
-		}
-		return
-	}
-
-	valAccount.ValidationError = valAccount.Account.Validate()
-
-	if pw.Password != pw.PasswordControl {
-		valAccount.FieldErrors["password"] = "Provided password did not match password control"
-		if valAccount.Message == "" {
-			valAccount.Message = valAccount.FieldErrors["password"]
-		}
-	}
-	if pw.Password == "" || pw.PasswordControl == "" {
-		valAccount.FieldErrors["password"] = "Please enter password and password control"
-		if valAccount.Message == "" {
-			valAccount.Message = valAccount.FieldErrors["password"]
-		}
-	}
-	if len(pw.Password) > 512 || len(pw.PasswordControl) > 512 {
-		valAccount.FieldErrors["password"] =
-			fmt.Sprintf("Entry too long, please shorten to %d characters", 512)
-		if valAccount.Message == "" {
-			valAccount.Message = valAccount.FieldErrors["password"]
-		}
-	}
-
-	captchaRes := r.PostForm.Get("captcha_resolve")
-	captchaId := r.PostForm.Get("captcha_id")
-	isCorrectCaptcha := rh.verifyCaptcha(captchaId, captchaRes)
-	if !isCorrectCaptcha {
-		valAccount.FieldErrors["password"] = "Please enter password and password control"
-		valAccount.FieldErrors["captcha"] = "Please resolve verification"
-		if valAccount.Message == "" {
-			valAccount.Message = valAccount.FieldErrors["captcha"]
-		}
-	}
-
-	if valAccount.Message != "" {
-		valAccount.CaptchaId = captcha.New()
-		err := tmpl.ExecuteTemplate(w, "layout", valAccount)
-		if err != nil {
-			panic(err)
-		}
-		return
-	}
-
-	valAccount.Account.SetPassword(pw.Password)
-	valAccount.Account.ActivationCode = sql.NullString{String: util.RandomToken(), Valid: true}
-
-	err = account.Create()
-	if err != nil {
-		valAccount.Message = "An error occurred during registration."
-		err := tmpl.ExecuteTemplate(w, "layout", valAccount)
-		if err != nil {
-			panic(err)
-		}
-		return
-	}
-
-	tmplFields := &struct {
-		From    string
-		To      string
-		Subject string
-		BaseUrl string
-		Code    string
-	}{}
-	tmplFields.From = conf.GetSmtpCredentials().From
-	tmplFields.To = account.Email
-	tmplFields.Subject = "GIN account activation"
-	tmplFields.BaseUrl = conf.GetServerConfig().BaseURL
-	tmplFields.Code = account.ActivationCode.String
-
-	content := util.MakeEmailTemplate("emailactivate.txt", tmplFields)
-	disp := util.NewEmailDispatcher()
-
-	err = disp.Send([]string{account.Email}, content.Bytes())
-	if err != nil {
-		msg := "An error occurred trying to send registration e-mail. Please contact an administrator."
-		PrintErrorHTML(w, r, msg, http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Add("Cache-Control", "no-store")
-	http.Redirect(w, r, "/oauth/registered_page", http.StatusFound)
-}
-
-// RegisteredPage displays information about how a newly created gin account can be activated.
-func RegisteredPage(w http.ResponseWriter, r *http.Request) {
-	head := "Account registered"
-	message := "Your account activation is pending. "
-	message += "An e-mail with an activation code has been sent to your e-mail address."
-
-	info := struct {
-		Header  string
-		Message string
-	}{head, message}
-
-	tmpl := conf.MakeTemplate("success.html")
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-	err := tmpl.ExecuteTemplate(w, "layout", info)
-	if err != nil {
-		panic(err)
-	}
-}
-
-// Activation removes an existing activation code from an account, thus rendering the account active.
-func Activation(w http.ResponseWriter, r *http.Request) {
-	err := r.ParseForm()
-	if err != nil {
-		PrintErrorHTML(w, r, "Activation request was malformed", http.StatusBadRequest)
-		return
-	}
-
-	getCode := r.Form.Get("activation_code")
-	if getCode == "" {
-		PrintErrorHTML(w, r, "Account activation code was absent", http.StatusBadRequest)
-		return
-	}
-
-	account, exists := data.GetAccountByActivationCode(getCode)
-	if !exists {
-		PrintErrorHTML(w, r, "Requested account does not exist", http.StatusNotFound)
-		return
-	}
-
-	account.ActivationCode.Valid = false
-	err = account.Update()
-	if err != nil {
-		panic(err)
-	}
-
-	head := "Account activation"
-	message := fmt.Sprintf("Congratulation %s %s! The account for %s has been activated and can now be used.",
-		account.FirstName, account.LastName, account.Login)
-	info := struct {
-		Header  string
-		Message string
-	}{head, message}
-
-	tmpl := conf.MakeTemplate("success.html")
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-
-	err = tmpl.ExecuteTemplate(w, "layout", info)
-	if err != nil {
-		panic(err)
-	}
-}
-
-type credentialData struct {
-	Credential string
-	ErrMessage string
-}
-
-// ResetInitPage provides an input form for resetting an account password
-func ResetInitPage(w http.ResponseWriter, r *http.Request) {
-
-	credData := &credentialData{}
-
-	tmpl := conf.MakeTemplate("resetinit.html")
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-	err := tmpl.ExecuteTemplate(w, "layout", credData)
-	if err != nil {
-		panic(err)
-	}
-}
-
-// ResetInit checks whether a provided login or e-mail address
-// belongs to a non-disabled account. If this is the case, the corresponding
-// account is updated with a password reset code and an email containing
-// the code is sent to the e-mail address of the account.
-func ResetInit(w http.ResponseWriter, r *http.Request) {
-
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-
-	credData := &credentialData{}
-
-	err := util.ReadFormIntoStruct(r, credData, true)
-	if err != nil {
-		panic(err)
-	}
-
-	if credData.Credential == "" {
-		credData.ErrMessage = "Please enter your login or e-mail address"
-		tmpl := conf.MakeTemplate("resetinit.html")
-		w.Header().Add("Warning", credData.ErrMessage)
-		err = tmpl.ExecuteTemplate(w, "layout", credData)
-		if err != nil {
-			panic(err)
-		}
-		return
-	}
-
-	account, ok := data.SetPasswordReset(credData.Credential)
-	if !ok {
-		credData.ErrMessage = "Invalid login or e-mail address"
-		tmpl := conf.MakeTemplate("resetinit.html")
-		w.Header().Add("Warning", credData.ErrMessage)
-		err = tmpl.ExecuteTemplate(w, "layout", credData)
-		if err != nil {
-			panic(err)
-		}
-		return
-	}
-
-	tmplFields := &struct {
-		From    string
-		To      string
-		Subject string
-		BaseUrl string
-		Code    string
-	}{}
-	tmplFields.From = conf.GetSmtpCredentials().From
-	tmplFields.To = account.Email
-	tmplFields.Subject = "Your GIN Account Password Reset Request"
-	tmplFields.BaseUrl = conf.GetServerConfig().BaseURL
-	tmplFields.Code = account.ResetPWCode.String
-
-	content := util.MakeEmailTemplate("emailreset.txt", tmplFields)
-	disp := util.NewEmailDispatcher()
-
-	err = disp.Send([]string{account.Email}, content.Bytes())
-	if err != nil {
-		msg := "An error occurred trying to send password reset e-mail. Please try again later."
-		PrintErrorHTML(w, r, msg, http.StatusInternalServerError)
-		return
-	}
-
-	head := "Success!"
-	message := "An e-mail with a password reset token has been sent to your e-mail address. "
-	message += "Please follow the contained link to reset your password. "
-	message += "Please note that your account will stay deactivated until your password reset has been completed."
-	info := struct {
-		Header  string
-		Message string
-	}{head, message}
-
-	tmpl := conf.MakeTemplate("success.html")
-	err = tmpl.ExecuteTemplate(w, "layout", info)
-}
-
-// ResetPage checks whether a password reset code submitted by request URI query exists and is still valid.
-// Display enter password form if valid, an error message otherwise.
-func ResetPage(w http.ResponseWriter, r *http.Request) {
-	err := r.ParseForm()
-	if err != nil {
-		PrintErrorHTML(w, r, "Request was malformed", http.StatusBadRequest)
-		return
-	}
-
-	code := r.Form.Get("reset_code")
-	if code == "" {
-		PrintErrorHTML(w, r, "Request was malformed", http.StatusBadRequest)
-		return
-	}
-
-	_, exists := data.GetAccountByResetPWCode(code)
-	if !exists {
-		PrintErrorHTML(w, r, "Your request is invalid or outdated. Please request a new reset code.",
-			http.StatusNotFound)
-		return
-	}
-
-	hidden := &struct {
-		ResetCode string
-		*util.ValidationError
-	}{code, &util.ValidationError{}}
-
-	tmpl := conf.MakeTemplate("reset.html")
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-	err = tmpl.ExecuteTemplate(w, "layout", hidden)
-}
-
-// Reset checks whether a submitted password reset code exists and is still valid. It further checks,
-// whether posted password and confirm password are identical and updates the account associated with
-// the password reset code with the new password. This update further removes any existing
-// password reset and account activation codes rendering the account active.
-func Reset(w http.ResponseWriter, r *http.Request) {
-	formData := &struct {
-		ResetCode       string
-		Password        string
-		PasswordControl string
-		*util.ValidationError
-	}{}
-
-	err := util.ReadFormIntoStruct(r, formData, true)
-	if err != nil {
-		panic(err)
-	}
-
-	account, exists := data.GetAccountByResetPWCode(formData.ResetCode)
-	if !exists {
-		PrintErrorHTML(w, r, "Your request is invalid or outdated. Please request a new reset code.",
-			http.StatusNotFound)
-		return
-	}
-
-	formData.ValidationError = &util.ValidationError{FieldErrors: make(map[string]string)}
-	if formData.Password != formData.PasswordControl {
-		formData.FieldErrors["password"] = "Provided password did not match password control"
-		formData.Message = formData.FieldErrors["password"]
-	}
-	if formData.Password == "" || formData.PasswordControl == "" {
-		formData.FieldErrors["password"] = "Please enter password and password control"
-		formData.Message = formData.FieldErrors["password"]
-	}
-	if len(formData.Password) > 512 || len(formData.PasswordControl) > 512 {
-		formData.FieldErrors["password"] =
-			fmt.Sprintf("Entry too long, please shorten to %d characters", 512)
-		formData.Message = formData.FieldErrors["password"]
-	}
-
-	if formData.FieldErrors["password"] != "" {
-		formData.Password = ""
-		formData.PasswordControl = ""
-		tmpl := conf.MakeTemplate("reset.html")
-		w.Header().Add("Cache-Control", "no-store")
-		w.Header().Add("Content-Type", "text/html")
-		w.Header().Add("Warning", formData.Message)
-		err := tmpl.ExecuteTemplate(w, "layout", formData)
-		if err != nil {
-			panic(err)
-		}
-		return
-	}
-
-	account.SetPassword(formData.Password)
-	account.ActivationCode.Valid = false
-	account.ResetPWCode.Valid = false
-	err = account.Update()
-	if err != nil {
-		panic(err)
-	}
-
-	head := "Your password has been reset!"
-	message := "You can now login using your new password."
-	info := struct {
-		Header  string
-		Message string
-	}{head, message}
-
-	tmpl := conf.MakeTemplate("success.html")
-	w.Header().Add("Cache-Control", "no-store")
-	w.Header().Add("Content-Type", "text/html")
-	err = tmpl.ExecuteTemplate(w, "layout", info)
 }
