+++ web/oauth_test.go
@@ -17,7 +17,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/G-Node/gin-auth/conf"
 	"github.com/G-Node/gin-auth/data"
 	"github.com/gorilla/mux"
 )
@@ -931,536 +930,5 @@ func TestValidate(t *testing.T) {
 	}
 	if result.Login != "alice" {
 		t.Errorf("Login expected to be 'alice' but was '%s'", result.Login)
-	}
-}
-
-func TestRegistrationPage(t *testing.T) {
-	handler := InitTestHttpHandler(t)
-
-	request, _ := http.NewRequest("GET", "/oauth/registration_page", strings.NewReader(""))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusOK {
-		t.Errorf("Response code '%d' expected but was '%d'", http.StatusOK, response.Code)
-	}
-}
-
-func TestRegistrationHandler(t *testing.T) {
-	data.InitTestDb(t)
-	f := func(id string, resolve string) bool {
-		return id != "" && id == resolve
-	}
-	handler := RegistrationHandler(f)
-
-	const registrationURL = "/oauth/registration"
-	const registeredPageURL = "/oauth/registered_page"
-
-	body := &url.Values{}
-	body.Add("Title", "Title")
-	body.Add("FirstName", "First Name")
-	body.Add("MiddleName", "Middle Name")
-	body.Add("LastName", "Last Name")
-	body.Add("Login", "tl")
-	body.Add("Email", "testemail@example.com")
-	body.Add("IsEmailPublic", "true")
-	body.Add("Institute", "Institute")
-	body.Add("Department", "Department")
-	body.Add("City", "City")
-	body.Add("Country", "Country")
-	body.Add("IsAffiliationPublic", "true")
-	body.Add("Password", "pw")
-	body.Add("PasswordControl", "pw")
-
-	// test that a request without a posted form does not redirect to another page
-	request, _ := http.NewRequest("POST", registrationURL, strings.NewReader(""))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	redirect, err := url.Parse(response.Header().Get("Location"))
-	if err != nil {
-		t.Error(err)
-	}
-	if redirect.String() != "" {
-		t.Errorf("Expected empty location header, but was '%s'", redirect.String())
-	}
-
-	// test that a request with correct form content but missing captcha stays on the same page
-	request, _ = http.NewRequest("POST", registrationURL, strings.NewReader(body.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	redirect, err = url.Parse(response.Header().Get("Location"))
-	if err != nil {
-		t.Error(err)
-	}
-	if redirect.String() != "" {
-		t.Errorf("Expected empty location header, but was '%s'", redirect.String())
-	}
-
-	// test that a request with correct form content but incorrect captcha stays on the same page
-	body.Add("captcha_id", "test")
-	request, _ = http.NewRequest("POST", registrationURL, strings.NewReader(body.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	redirect, err = url.Parse(response.Header().Get("Location"))
-	if err != nil {
-		t.Error(err)
-	}
-	if redirect.String() != "" {
-		t.Errorf("Expected empty location header, but was '%s'", redirect.String())
-	}
-
-	// test that a request with correct form content redirects to registered_page
-	body.Add("captcha_resolve", "test")
-	request, _ = http.NewRequest("POST", registrationURL, strings.NewReader(body.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	redirect, err = url.Parse(response.Header().Get("Location"))
-	if err != nil {
-		t.Error(err)
-	}
-	if redirect.String() != registeredPageURL {
-		t.Errorf("Expected to be redirected to '%s', but was '%s'", registeredPageURL, redirect.String())
-	}
-
-	// Test error when sending e-mail
-	body.Set("Login", "tl2")
-	body.Set("Email", "testemail2@example.com")
-
-	mode := conf.GetSmtpCredentials().Mode
-	host := conf.GetSmtpCredentials().Host
-	conf.GetSmtpCredentials().Mode = ""
-	conf.GetSmtpCredentials().Host = "iDoNotExist.com"
-	request, _ = http.NewRequest("POST", registrationURL, strings.NewReader(body.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusInternalServerError {
-		t.Errorf("Expected StatusInternatServerError but got '%d'", response.Code)
-	}
-
-	// reset smtp configuration
-	conf.GetSmtpCredentials().Mode = mode
-	conf.GetSmtpCredentials().Host = host
-}
-
-func TestRegisteredPage(t *testing.T) {
-	handler := InitTestHttpHandler(t)
-
-	request, _ := http.NewRequest("GET", "/oauth/registered_page", strings.NewReader(""))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusOK {
-		t.Errorf("Response code '%d' expected but was '%d'", http.StatusOK, response.Code)
-	}
-}
-
-func TestActivation(t *testing.T) {
-	handler := InitTestHttpHandler(t)
-	const activationURL = "/oauth/activation"
-	const activationCodeDisabled = "ac_b"
-	const activationCode = "ac_a"
-
-	// Test missing query
-	request, _ := http.NewRequest("GET", activationURL, strings.NewReader(""))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusBadRequest {
-		t.Errorf("Expected StatusBadRequest on empty activationCode but got '%d'", response.Code)
-	}
-
-	// Test invalid activation code
-	request, _ = http.NewRequest("GET", activationURL, strings.NewReader(""))
-	q := request.URL.Query()
-	q.Add("activation_code", "iDoNotExist")
-	request.URL.RawQuery = q.Encode()
-
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusNotFound {
-		t.Errorf("Expected StatusNotFound on invalid activationCode but got '%d'", response.Code)
-	}
-
-	// Test activation code of disabled account
-	request, _ = http.NewRequest("GET", activationURL, strings.NewReader(""))
-	q = request.URL.Query()
-	q.Add("activation_code", activationCodeDisabled)
-	request.URL.RawQuery = q.Encode()
-
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusNotFound {
-		t.Errorf("Expected StatusNotFound on disabled account but got '%d'", response.Code)
-	}
-
-	// Test valid activation
-	account, exists := data.GetAccountByActivationCode(activationCode)
-	if !exists {
-		t.Errorf("Error on fetching account by activation code '%s'", activationCode)
-	}
-	if account.ActivationCode.String != activationCode {
-		t.Errorf("Expected activation code to be '%s' but got '%s'",
-			activationCode, account.ActivationCode.String)
-	}
-	accountLogin := account.Login
-
-	request, _ = http.NewRequest("GET", activationURL, strings.NewReader(""))
-	q = request.URL.Query()
-	q.Add("activation_code", activationCode)
-	request.URL.RawQuery = q.Encode()
-
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK on valid activationCode but got '%d'", response.Code)
-	}
-
-	account, exists = data.GetAccountByLogin(accountLogin)
-	if !exists {
-		t.Errorf("Error on fetching account by login '%s'", accountLogin)
-	}
-	if account.ActivationCode.String != "" || account.ActivationCode.Valid {
-		t.Errorf("Activation code should be cleared after activation but was '%s', '%t'",
-			account.ActivationCode.String, account.ActivationCode.Valid)
-	}
-}
-
-func TestResetInitPage(t *testing.T) {
-	handler := InitTestHttpHandler(t)
-	const resetURL = "/oauth/reset_init_page"
-
-	request, _ := http.NewRequest("GET", resetURL, strings.NewReader(""))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK but got '%d'", response.Code)
-	}
-}
-
-func TestResetInit(t *testing.T) {
-	handler := InitTestHttpHandler(t)
-
-	const resetInitURL = "/oauth/reset_init"
-	const disabledLogin = "inact_log4"
-	const disabledEmail = "email4@example.com"
-	const enabledLogin = "inact_log1"
-
-	// Test post empty body
-	request, _ := http.NewRequest("POST", resetInitURL, strings.NewReader(""))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "Please enter your login or e-mail address" {
-		t.Errorf("Expected empty id field message but got '%s'", response.Header().Get("Warning"))
-	}
-
-	mkBody := &url.Values{}
-	mkBody.Add("Credential", "iDoNotExist")
-
-	// Test invalid login
-	request, _ = http.NewRequest("POST", resetInitURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "Invalid login or e-mail address" {
-		t.Errorf("Expected invalid login message but got '%s'", response.Header().Get("Warning"))
-	}
-
-	// Test login of disabled account
-	mkBody = &url.Values{}
-	mkBody.Add("Credential", disabledLogin)
-
-	request, _ = http.NewRequest("POST", resetInitURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "Invalid login or e-mail address" {
-		t.Errorf("Expected invalid login message but got '%s'", response.Header().Get("Warning"))
-	}
-
-	// Test e-mail of disabled account
-	mkBody = &url.Values{}
-	mkBody.Add("Credential", disabledEmail)
-
-	request, _ = http.NewRequest("POST", resetInitURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "Invalid login or e-mail address" {
-		t.Errorf("Expected invalid login message but got '%s'", response.Header().Get("Warning"))
-	}
-
-	// Test valid update using login
-	mkBody = &url.Values{}
-	mkBody.Add("Credential", enabledLogin)
-
-	request, _ = http.NewRequest("POST", resetInitURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "" {
-		t.Errorf("Expected empty warning message but got '%s'", response.Header().Get("Warning"))
-	}
-
-	// Test error when sending e-mail
-	mode := conf.GetSmtpCredentials().Mode
-	host := conf.GetSmtpCredentials().Host
-	conf.GetSmtpCredentials().Mode = ""
-	conf.GetSmtpCredentials().Host = "iDoNotExist.com"
-	request, _ = http.NewRequest("POST", resetInitURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusInternalServerError {
-		t.Errorf("Expected StatusInternalServerError but got '%d'", response.Code)
-	}
-
-	// reset smtp configuration
-	conf.GetSmtpCredentials().Mode = mode
-	conf.GetSmtpCredentials().Host = host
-}
-
-func TestResetPage(t *testing.T) {
-	handler := InitTestHttpHandler(t)
-	const resetURL = "/oauth/reset_page"
-	const codeKey = "reset_code"
-	const codeInvalid = "iDoNotExist"
-	const codeDisabled = "rc_c"
-	const codeValid = "rc_a"
-
-	// Test missing password reset code
-	request, _ := http.NewRequest("GET", resetURL, strings.NewReader(""))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusBadRequest {
-		t.Errorf("Expected StatusBadRequest on missing reset code but got '%d'", response.Code)
-	}
-
-	// Test invalid password reset code
-	request, _ = http.NewRequest("GET", resetURL, strings.NewReader(""))
-	q := request.URL.Query()
-	q.Add(codeKey, codeInvalid)
-	request.URL.RawQuery = q.Encode()
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusNotFound {
-		t.Errorf("Expected StatusNotFound on invalid reset code but got '%d'", response.Code)
-	}
-
-	// Test valid password reset code of disabled account
-	request, _ = http.NewRequest("GET", resetURL, strings.NewReader(""))
-	q = request.URL.Query()
-	q.Add(codeKey, codeDisabled)
-	request.URL.RawQuery = q.Encode()
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusNotFound {
-		t.Errorf("Expected StatusNotFound on disabled reset code but got '%d'", response.Code)
-	}
-
-	// Test valid password reset code
-	request, _ = http.NewRequest("GET", resetURL, strings.NewReader(""))
-	q = request.URL.Query()
-	q.Add(codeKey, codeValid)
-	request.URL.RawQuery = q.Encode()
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK on valid reset code but got '%d'", response.Code)
-	}
-}
-
-func TestReset(t *testing.T) {
-	handler := InitTestHttpHandler(t)
-	const resetURL = "/oauth/reset"
-	const codeKey = "ResetCode"
-	const codeInvalid = "iDoNotExist"
-	const codeDisabled = "rc_c"
-	const codeValid = "rc_a"
-	const codeValidInactive = "rc_b"
-
-	// Test empty body
-	request, _ := http.NewRequest("POST", resetURL, strings.NewReader(""))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusNotFound {
-		t.Errorf("Expected StatusNotFound on missing reset code but got '%d'", response.Code)
-	}
-
-	// Test invalid password reset code
-	mkBody := &url.Values{}
-	mkBody.Add(codeKey, codeInvalid)
-
-	request, _ = http.NewRequest("POST", resetURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusNotFound {
-		t.Errorf("Expected StatusNotFound on invalid reset code but got '%d'", response.Code)
-	}
-
-	// Test valid password reset code of disabled account
-	mkBody.Set(codeKey, codeDisabled)
-
-	request, _ = http.NewRequest("POST", resetURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusNotFound {
-		t.Errorf("Expected StatusNotFound on disabled reset code but got '%d'", response.Code)
-	}
-
-	// Test valid password reset code, missing password
-	mkBody.Set(codeKey, codeValid)
-	request, _ = http.NewRequest("POST", resetURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK on valid reset code but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "Please enter password and password control" {
-		t.Errorf("Expected empty password warning but got '%s'", response.Header().Get("Warning"))
-	}
-
-	// Test valid password reset code, password and password control missmatch
-	mkBody.Add("Password", "pw")
-	mkBody.Add("PasswordControl", "pwcontrol")
-	request, _ = http.NewRequest("POST", resetURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK on valid reset code but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "Provided password did not match password control" {
-		t.Errorf("Expected empty password warning but got '%s'", response.Header().Get("Warning"))
-	}
-
-	// Test valid password reset code, password too long
-	s := []string{}
-	for i := 0; i < 513; i++ {
-		s = append(s, "s")
-	}
-	js := strings.Join(s, "")
-
-	mkBody.Set("Password", js)
-	request, _ = http.NewRequest("POST", resetURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK on valid reset code but got '%d'", response.Code)
-	}
-	if response.Header().Get("Warning") != "Entry too long, please shorten to 512 characters" {
-		t.Errorf("Expected empty password warning but got '%s'", response.Header().Get("Warning"))
-	}
-
-	// Test valid password reset code, reset of pw code
-	account, exists := data.GetAccountByResetPWCode(codeValid)
-	if !exists {
-		t.Errorf("Account with reset code '%s' does not exist", codeValid)
-	}
-	id := account.UUID
-	_, exists = data.GetAccount(id)
-	if exists {
-		t.Errorf("Account with id '%s' reset code '%s' should not be active", id, codeValid)
-	}
-	pwHash := account.PWHash
-
-	mkBody.Set("Password", "pw")
-	mkBody.Set("PasswordControl", "pw")
-	request, _ = http.NewRequest("POST", resetURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK on valid reset code but got '%d'", response.Code)
-	}
-
-	account, exists = data.GetAccount(id)
-	if !exists {
-		t.Errorf("Account with id '%s' should be active", id)
-	}
-	if account.PWHash == pwHash {
-		t.Errorf("Password of Account with id '%s' has not been updated", id)
-	}
-	if account.ResetPWCode.String != "" {
-		t.Errorf("Password reset code of Account with id '%s' has not been deleted", id)
-	}
-
-	// Test valid password reset code, reset of pw code, reset of activation code
-	account, exists = data.GetAccountByResetPWCode(codeValidInactive)
-	if !exists {
-		t.Errorf("Account with reset code '%s' does not exist", codeValidInactive)
-	}
-	id = account.UUID
-	_, exists = data.GetAccount(id)
-	if exists {
-		t.Errorf("Account with id '%s' reset code '%s' should not be active", id, codeValidInactive)
-	}
-	pwHash = account.PWHash
-
-	mkBody.Set(codeKey, codeValidInactive)
-	request, _ = http.NewRequest("POST", resetURL, strings.NewReader(mkBody.Encode()))
-	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	response = httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Errorf("Expected StatusOK on valid reset code but got '%d'", response.Code)
-	}
-
-	account, exists = data.GetAccount(id)
-	if !exists {
-		t.Errorf("Account with id '%s' should be active", id)
-	}
-	if account.PWHash == pwHash {
-		t.Errorf("Password of Account with id '%s' has not been updated", id)
-	}
-	if account.ResetPWCode.String != "" {
-		t.Errorf("Password reset code of Account with id '%s' has not been removed", id)
-	}
-	if account.ActivationCode.String != "" {
-		t.Errorf("Activation code of Account with id '%s' has not been removed", id)
 	}
 }
